scripts/requirements.txt
balldontlie
python-dotenv
pymongo 
dnspython
certifi